requirements-dsp.txt
dspy-ai>=2.3.0
openai>=1.0.0
anthropic>=0.5.0
promptfoo>=0.25.0
tiktoken>=0.5.0
pyyaml>=6.0
colorlog>=6.7.0
tqdm>=4.66.0
pytest>=7.3.1
pygments>=2.15.0
difflib36>=0.3.0
gitpython>=3.1.40
rich>=13.4.2
sqlite-utils>=3.35 